requirements.txt
alabaster==0.7.12
anyjson==0.3.3
appdirs==1.4.3
astroid==2.0.4
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-datetime-utc==1.0.2
django-filter==2.0.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2.post0
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.9.0
idna==2.7
imagesize==1.1.0
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
packaging==18.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
slugify==0.0.1
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
uWSGI==2.0.17.1
validate-email==1.3
whitenoise==4.1
wrapt==1.10.11
